scripts/lending_tower/requirements.txt
pymysql==1.1.0
pymssql==2.2.11
boto3==1.34.0
python-dotenv==1.0.0
